requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-api-tracking==1.8.4
filelock==3.14.0
numpy==2.0.0
pandas==2.2.2
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.2